helper script for UPI on PowerVS Private Cloud
Initial commit. Needs some cleanup.

This script is meant to be used with the helper node playbook to
provide: load balancing, DHCP, and DNS.

Some of the values are sourced from the helper node vars file
using yq so as not to require users to sync two sources of
truth.

Signed-off-by: Christy Norman <christy@linux.vnet.ibm.com>